+++ packages/apollo-utilities/CHANGELOG.md
@@ -24,3 +24,4 @@
 - Added removeDirectivesFromDocument to allow cleaning of client only directives
 - Added hasDirectives to recurse the AST and return a boolean for an array of directive names
 - Improved performance of common store actions by memoizing addTypename and removeConnectionDirective
+- Fix: Do not add `__typename` only to root mutations, add to root queries